refactor(home_cubit): Move home cubit to state_management folder
